requirements.txt
flask==2.0.1
flask-cors==3.0.10
sqlalchemy==1.4.23
openai==1.3.0
python-dotenv==0.19.0
requests==2.26.0
google-generativeai==0.3.2
ctransformers==0.2.27
sentencepiece==0.1.99 